pr comments: moving packages, updated lint, nits
