+++ auth/consts.go
@@ -11,6 +11,10 @@ const (
 	ED25519ID   uint8 = 0
 	SECP256R1ID uint8 = 1
 	BLSID       uint8 = 2
+
+	ED25519Key   = "ed25519"
+	Secp256r1Key = "secp256r1"
+	BLSKey       = "bls"
 )
 
 func Engines() map[uint8]vm.AuthEngine {
