+++ examples/morpheusvm/auth/keys.go
@@ -1,6 +1,10 @@
 // Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
+// Package auth provides utilities for generating and loading private keys.
+// This package is only used for testing and CLI purposes and is not required
+// to be implemented by the VM developer.
+
 package auth
 
 import (
@@ -14,17 +18,14 @@ import (
 	"github.com/ava-labs/hypersdk/utils"
 )
 
-const (
-	ed25519Key   = "ed25519"
-	secp256r1Key = "secp256r1"
-	blsKey       = "bls"
-)
-
 var ErrInvalidKeyType = errors.New("invalid key type")
 
+// TODO: make these functions general purpose where the VM provides a set of valid strings,
+// functions to generate corresponding new private keys, and the functionality
+// to unmarshal private key bytes into the correct type.
 func CheckKeyType(k string) error {
 	switch k {
-	case ed25519Key, secp256r1Key, blsKey:
+	case auth.ED25519Key, auth.Secp256r1Key, auth.BLSKey:
 		return nil
 	default:
 		return fmt.Errorf("%w: %s", ErrInvalidKeyType, k)
@@ -33,7 +34,7 @@ func CheckKeyType(k string) error {
 
 func GeneratePrivateKey(k string) (*auth.PrivateKey, error) {
 	switch k {
-	case ed25519Key:
+	case auth.ED25519Key:
 		p, err := ed25519.GeneratePrivateKey()
 		if err != nil {
 			return nil, err
@@ -42,7 +43,7 @@ func GeneratePrivateKey(k string) (*auth.PrivateKey, error) {
 			Address: auth.NewED25519Address(p.PublicKey()),
 			Bytes:   p[:],
 		}, nil
-	case secp256r1Key:
+	case auth.Secp256r1Key:
 		p, err := secp256r1.GeneratePrivateKey()
 		if err != nil {
 			return nil, err
@@ -51,7 +52,7 @@ func GeneratePrivateKey(k string) (*auth.PrivateKey, error) {
 			Address: auth.NewSECP256R1Address(p.PublicKey()),
 			Bytes:   p[:],
 		}, nil
-	case blsKey:
+	case auth.BLSKey:
 		p, err := bls.GeneratePrivateKey()
 		if err != nil {
 			return nil, err
@@ -67,7 +68,7 @@ func GeneratePrivateKey(k string) (*auth.PrivateKey, error) {
 
 func LoadPrivateKey(k string, path string) (*auth.PrivateKey, error) {
 	switch k {
-	case ed25519Key:
+	case auth.ED25519Key:
 		p, err := utils.LoadBytes(path, ed25519.PrivateKeyLen)
 		if err != nil {
 			return nil, err
@@ -77,7 +78,7 @@ func LoadPrivateKey(k string, path string) (*auth.PrivateKey, error) {
 			Address: auth.NewED25519Address(pk.PublicKey()),
 			Bytes:   p,
 		}, nil
-	case secp256r1Key:
+	case auth.Secp256r1Key:
 		p, err := utils.LoadBytes(path, secp256r1.PrivateKeyLen)
 		if err != nil {
 			return nil, err
@@ -87,7 +88,7 @@ func LoadPrivateKey(k string, path string) (*auth.PrivateKey, error) {
 			Address: auth.NewSECP256R1Address(pk.PublicKey()),
 			Bytes:   p,
 		}, nil
-	case blsKey:
+	case auth.BLSKey:
 		p, err := utils.LoadBytes(path, bls.PrivateKeyLen)
 		if err != nil {
 			return nil, err
