+++ tests/e2e/e2e.go
@@ -117,6 +117,10 @@ var _ = ginkgo.Describe("[HyperSDK Tx Workloads]", func() {
 
 var _ = ginkgo.Describe("[HyperSDK Spam Workloads]", func() {
 	ginkgo.It("Spam Workload", func() {
+		if spamKey == nil || spamHelper == nil {
+			return
+		}
+
 		tc := e2e.NewTestContext()
 		require := require.New(tc)
 		blockchainID := e2e.GetEnv(tc).GetNetwork().GetSubnet(vmName).Chains[0].ChainID
